requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu130
streamlit
opencv-python
pillow
torch
torchvision
transformers>=4.51.0
accelerate
numpy
pgvector
psycopg2-binary
python-dotenv
sentence-transformers
